+++ CHANGELOG.md
@@ -57,6 +57,9 @@ All notable changes to this project will be documented in this file, in reverse 
   can be now used with PHP 7.1+ instead of `session.hash_bits_per_character`
   (used with PHP versions prior to 7.1).
 
+- [#118](https://github.com/zendframework/zend-session/pull/118) avoid unnecessary phpinfo() call
+  when register own save handler which is an object.
+
 ## 2.8.5 - 2018-02-22
 
 ### Added